pyproject.toml

[tool.poetry]
name = "sm2t_api"
version = "0.1.0"
description = "API serving traffic information for OSM highways"
authors = ["Christina Ludwig <christina.ludwig@uni-heidelberg.de>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0"
geopandas = "^0.9"
Flask-RESTful = "^0.3"
psycopg2 = "^2.9"
SQLAlchemy = "^1.4.29"
GeoAlchemy2 = "^0.10.2"
python-dotenv = "^0.19.2"
uWSGI = "^2.0.20"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pre-commit = "^2.16.0"
requests = "^2.27.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
